Rename Plotting subdirectory to Common
There are other similar interfaces which will be added to this directory.

Re #23027
