+++ qt/scientific_interfaces/ISISReflectometry/GUI/MainWindow/MainWindowView.cpp
@@ -7,7 +7,7 @@
 #include "MainWindowView.h"
 #include "Common/IndexOf.h"
 #include "GUI/Batch/BatchView.h"
-#include "GUI/Plotting/Plotter.h"
+#include "GUI/Common/Plotter.h"
 #include "GUI/Runs/CatalogRunNotifier.h"
 #include "MantidKernel/make_unique.h"
 #include <QMessageBox>
